pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = "Contacts API"
authors = ["Lesia Soloviova <solovyova.lesya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.37"}
psycopg2 = "^2.9.10"
pydantic = "^2.10.5"
alembic = "^1.14.0"
pydantic-settings = "^2.7.1"
asyncpg = "^0.30.0"
email-validator = "^2.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
python-multipart = "^0.0.20"
fastapi-mail = "^1.4.2"
slowapi = "^0.1.9"
cloudinary = "^1.42.1"
sphinx = "^8.1.3"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"
aiosqlite = "^0.20.0"
pytest-cov = "^6.0.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"
asyncio_default_fixture_loop_scope = "function"
